Restructures checkin eligibility with new service function
