+++ controller/controller.go
@@ -59,18 +59,18 @@ func CreateUserCheckin(w http.ResponseWriter, r *http.Request) {
 	json.NewDecoder(r.Body).Decode(&user_checkin)
 
 	is_user_registered, err := service.IsUserRegistered(user_checkin.ID)
-	
+
 	if err != nil {
 		panic(errors.UnprocessableError(err.Error()))
 	}
-	
+
 	user_checkin_allowed := false
 
 	// To checkin, the user must either (have RSVPed) or (have registered and got an override)
 	if is_user_registered && user_checkin.Override {
 
 		user_checkin_allowed = true
-		
+
 	} else {
 
 		isRsvped, err := service.IsAttendeeRsvped(user_checkin.ID)
